requirements.txt
Django
psycopg2_binary
djangorestframework
django-filter
python-dotenv

